Move hooks from data to a single hook file.
