+++ src/data/faqEntry/helper.ts
@@ -1,0 +1,9 @@
+import * as R from 'ramda'
+import { FAQPage, FAQEntry } from '.'
+
+export const hasFAQEntries: (faqPage: FAQPage) => boolean = R.pipe<
+  FAQPage,
+  FAQEntry[],
+  number,
+  boolean
+>(R.prop('faqEntries'), R.length, R.lte(1))
