+++ src/data/faqEntry/index.tsx
@@ -1,6 +1,4 @@
-import * as R from 'ramda'
 import { Locale, TranslationCollection, filterByLocale } from '../i18n'
-import { useLocale } from '../../utils/hooks'
 import data from './faq.json'
 import { Head } from '../../components/Head'
 
@@ -29,12 +27,3 @@ export interface FAQEntry {
 
 export const getFAQPage = (locale: Locale): FAQPage =>
   filterByLocale(locale)(faq)
-
-export const useFAQPage: () => FAQPage = R.pipe(useLocale, getFAQPage)
-
-export const hasFAQEntries: (faqPage: FAQPage) => boolean = R.pipe<
-  FAQPage,
-  FAQEntry[],
-  number,
-  boolean
->(R.prop('faqEntries'), R.length, R.lte(1))
