+++ src/data/footer/index.tsx
@@ -1,6 +1,4 @@
-import * as R from 'ramda'
 import data from './footer.json'
-import { useLocale } from '../../utils/hooks'
 import { translate, Locale, TranslationCollection } from '../../cms/i18n'
 
 export interface FooterData {
@@ -20,6 +18,6 @@ export interface FooterDataOnDisk {
 // TODO: add type analysis of imported data
 //       checkout https://github.com/pelotom/runtypes
 const typecheckedData: FooterDataOnDisk = data
-const getFooter = (locale: Locale): FooterData =>
+
+export const getFooter = (locale: Locale): FooterData =>
   translate(locale)(typecheckedData)
-export const useFooter = R.pipe(useLocale, getFooter)
