+++ src/data/imprint/index.tsx
@@ -1,6 +1,4 @@
-import * as R from 'ramda'
 import { Locale, TranslationCollection, filterByLocale } from '../i18n'
-import { useLocale } from '../../utils/hooks'
 import data from './imprint.json'
 import { Head } from '../../components/Head'
 
@@ -24,5 +22,3 @@ const imprint: ImprintPageOnDisk = data
 
 export const getImprint = (locale: Locale): Imprint =>
   filterByLocale(locale)(imprint)
-
-export const useImprint: () => Imprint = R.pipe(useLocale, getImprint)
