+++ src/data/navbar/index.tsx
@@ -1,6 +1,4 @@
-import * as R from 'ramda'
 import { Locale, TranslationCollection, filterByLocale } from '../i18n'
-import { useLocale } from '../../utils/hooks'
 import data from './navbar.json'
 
 export interface Navbar {
@@ -10,6 +8,5 @@ export interface Navbar {
 
 const navbar: TranslationCollection<Navbar> = data
 
-const getNavbar = (locale: Locale): Navbar => filterByLocale(locale)(navbar)
-
-export const useNavbar: () => Navbar = R.pipe(useLocale, getNavbar)
+export const getNavbar = (locale: Locale): Navbar =>
+  filterByLocale(locale)(navbar)
